+++ spring/spring-batch/fastcampus/src/main/java/wh/duckbill/fastcampus/part3/TaskAndChunk/TaskJobParametersProcessingConfiguration.java
@@ -56,7 +56,7 @@ public class TaskJobParametersProcessingConfiguration {
 
             String value = jobParameters.getString("chunkSize", "10");
             int chunkSize = StringUtils.isNotEmpty(value) ? Integer.parseInt(value) : 10;
-            int fromIndex = stepExecution.getReadCount();
+            int fromIndex = (int) stepExecution.getReadCount();
             int toIndex = fromIndex + chunkSize;
             if (toIndex > items.size())
                 toIndex = items.size();
